build(circular-deps): tsPreCompilationDeps only fixes for gif and video
