+++ packages/react-components/src/components/video/video-overlay.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled'
 import React, { useEffect, useState } from 'react'
-import { GifOverlayProps } from '../gif'
+import { GifOverlayProps } from '../types'
 import Video from './'
 import { VolumeOffIcon, VolumeOnIcon } from './controls/volume'
