+++ packages/react-components/src/components/video/video.tsx
@@ -2,9 +2,8 @@ import { pingback } from '@giphy/js-analytics'
 import { IGif, IImage } from '@giphy/js-types'
 import { getBestVideo, getGifHeight } from '@giphy/js-util'
 import React, { useCallback, useEffect, useRef, useState } from 'react'
-import { getErrorMessage, shouldFireQuartile } from './util'
+import { getErrorMessage, QuartileEvent, shouldFireQuartile } from './util'
 
-export type QuartileEvent = 0.25 | 0.5 | 0.75
 const quartileEvents: QuartileEvent[] = [0.25, 0.5, 0.75]
 export type MEDIA_STATE = 'playing' | 'paused'
 const videoClassName = 'giphy-video'
