+++ packages/react-components/src/index.tsx
@@ -5,12 +5,12 @@ export { default as Attribution } from './components/attribution'
 export { default as AttributionOverlay } from './components/attribution/overlay'
 export { default as Carousel } from './components/carousel'
 export { default as Gif } from './components/gif'
-export type { GifOverlayProps } from './components/gif'
 export { default as Grid } from './components/grid'
 export { PingbackContext } from './components/pingback-context-manager'
 export { default as SearchBar } from './components/search-bar'
 export { default as SearchContextManager, SearchContext } from './components/search-bar/context'
 export { default as SuggestionBar } from './components/search-bar/suggestion-bar'
+export type { GifOverlayProps } from './components/types'
 export { default as Video } from './components/video'
 export * from './components/video/controls/play-pause'
 export * from './components/video/controls/volume'
